backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
qrcode==8.0
Pillow==11.1.0
mysqlclient==2.2.7
python-decouple==3.8
djangorestframework-simplejwt==5.3.1

PyPDF2==3.0.0
reportlab==4.0.0
